backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
Werkzeug==3.0.1
marshmallow==3.20.1
requests==2.31.0
# 推荐系统依赖
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
jieba==0.42.1
scipy==1.11.4
APScheduler==3.10.4
